LayoutTests/imported/w3c/web-platform-tests/html/browsers/windows/nested-browsing-contexts/window-parent-null-expected.txt


FAIL `window.parent` is null when browsing context container element removed assert_equals: expected null but got object "[object Window]"
FAIL `window.parent` null when parent browsing context container removed assert_equals: expected null but got object "[object Window]"
